6.44

+++ src/main/java/com/sportshop/sports_shop/repository/BienTheSanPhamRepository.java
@@ -9,5 +9,5 @@ import com.sportshop.sports_shop.model.BienTheSanPham;
 
 @Repository
 public interface BienTheSanPhamRepository extends JpaRepository<BienTheSanPham, Integer> {
-    List<BienTheSanPham> findBySanPhamMaSp(Integer maSp);
+    List<BienTheSanPham> findBySanPhamMaSp(Long maSp);
 }
